+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change log
 
 For detailed information check [Blogophon's releases on Github](https://github.com/fboes/blogophon/releases).
 
+* 💊 Fix ARIA handling of gallery links
 * 🎁 Add Gopher server support
 * 🎁 Add Spotify embedding
 * 🎁 JSON-LD added